refactor(gateway): stage 4b slices + empty allowlist — ironclaw#2599
Collapses the last pre-existing back-edges tracked by the boundary
checker and moves the first three small feature slices out of
server.rs.

Stage 4b slices:

- features/logs/       — /api/logs/events, /api/logs/level (GET/PUT)
- features/pairing/    — /api/pairing/{channel} (GET),
                         /api/pairing/{channel}/approve (POST)
- features/status/     — /api/gateway/status (+ GatewayStatusResponse,
                         ModelUsageEntry, each now owned by the slice)

Platform extensions (co-located with existing platform modules so
every caller — handlers/, features/, and the still-shrinking
server.rs — can reach them without a back-edge):

- platform/legacy_auth.rs:
  - handle_legacy_auth_token_submission
  - handle_legacy_auth_cancel
  - clear_auth_mode, clear_auth_mode_for_thread
  Consumers: server.rs chat HTTP shims + platform/ws.rs.
- platform/engine_dispatch.rs:
  - dispatch_engine_submission
  - dispatch_engine_external_callback
  - dispatch_onboarding_ready_followup  (now takes &ExtensionName)
  Consumers: server.rs chat + extensions_setup_submit + features/pairing.
- platform/static_files.rs gains the workspace-backed layout/widget
  readers (read_layout_config, load_resolved_widgets,
  read_widget_manifest, LAYOUT_PATH, WIDGETS_DIR, MAX_WIDGET_* caps).
  handlers/frontend.rs imports them back from platform.

ExtensionName adoption:

- Deletes sanitize_extension_name and its 5 unit tests from
  server.rs. The defensive "never fails, returns 'unknown'" helper is
  replaced with ironclaw_common::ExtensionName validation at the one
  untrusted boundary we still expose (pairing_approve_handler's URL
  path). Invalid names now return 400 at the handler — the old
  behavior sanitized injection-shaped input into a safe-but-nonsense
  string that would never have matched a real extension anyway, so
  this is strictly better telemetry with no loss of reachable
  behavior. Registry-sourced names (derive_onboarding in
  handlers/extensions.rs) drop the sanitize call entirely; the
  comment notes a follow-up to type Extension.name as ExtensionName
  directly.
- Other shared helpers that needed to leave server.rs as collateral:
  images_to_attachments moves to web/util.rs alongside the other
  pure message-building helpers.

ws.rs cleanup:

- Switches GatewayState / PerUserRateLimiter / RateLimiter /
  ActiveConfigSnapshot imports from the server.rs re-export path to
  crate::channels::web::platform::state directly, removing the last
  state-type allowlist entries.

Allowlist:

- scripts/check_gateway_boundaries.py: ALLOWLIST is now empty. All
  eight pre-existing entries (widget helpers, seven ws.rs shim
  symbols) are gone — every relocation landed in platform/. The
  allowlist mechanism stays in place for future narrowly-scoped
  exceptions.

Diff is roughly −700 lines net from server.rs (now ~5,700 down
from ~6,300), spread across three new feature-slice files and two
new platform modules. No behavior change — this is a pure
relocation + one type-boundary upgrade.
